Do giant giant crab earlier
