+++ packages/garbo/src/potions.ts
@@ -57,6 +57,7 @@ import {
   baseMeat,
   bestShadowRift,
   HIGHLIGHT,
+  improvesAStat,
   pillkeeperOpportunityCost,
   targetingMeat,
   targetMeat,
@@ -678,6 +679,12 @@ export const farmingPotions = [
   ...(have($item`closed-circuit pay phone`) ? [rufusPotion] : []),
 ];
 
+export function getFarmingPotions(avoidStats = false): Potion[] {
+  return avoidStats
+    ? farmingPotions.filter((potion) => !improvesAStat(potion.effect()))
+    : farmingPotions;
+}
+
 export function doublingPotions(targets: number): Potion[] {
   return farmingPotions
     .filter(
@@ -711,7 +718,7 @@ export function potionSetupCompleted(): boolean {
  * Determines if potions are worth using by comparing against meat-equilibrium. Considers using pillkeeper to double them. Accounts for non-wanderer targets. Does not account for PYEC/LTC, or running out of turns with the ascend flag.
  * @param targetsOnly Are we valuing the potions only for meat targets (noBarf)?
  */
-export function potionSetup(targetsOnly: boolean): void {
+export function potionSetup(targetsOnly: boolean, avoidStats = false): void {
   castAugustScepterBuffs();
   // TODO: Count PYEC.
   // TODO: Count free fights (25 meat each for most).
@@ -746,7 +753,7 @@ export function potionSetup(targetsOnly: boolean): void {
     }
 
     // Only test potions which are reasonably close to being profitable using historical price.
-    const testPotions = farmingPotions.filter(
+    const testPotions = getFarmingPotions(avoidStats).filter(
       (potion) => potion.gross(targets) / potion.price(true) > 0.5,
     );
     const nonWishTestPotions = testPotions.filter(
