+++ packages/garbo/src/tasks/cockroach/index.ts
@@ -1,2 +1,1 @@
 export { CockroachSetup } from "./prep";
-export { CockroachFinish } from "./finish";
